Added notification support
Still very raw and undercooked, will have to work on it in the future.
whats working:
there can only be 3 popup notifications at any time, will have to add it
to the config
any notifications that are added when there's no space left will be
added to a queue
right now we just don't care about the reason why a notification is
dismissed

will have to fix the styling and edit the style of the pill to "merge"
the two together.
will also have to hide the notifications and pause for a bit when the
pill is expanded, and maybe reflash the notifications again once it's
un-expanded.
must add a "Caffeine" tile to pause the notifications, will have to
decide what would happen after unpausing the notifications: a- flash all
missed notifications b- just don't
will add serialization and notification history later when i add the
notification center applet