Made the pill window an overlay
the pill is now an overlay, meaning it'll still be visible when a window
is fullscreen.
the pill will hide itself when a client's fullscreen state == 2, meaning
it occupies the entire screen.
the pill will reveal itself only when an applet is maximized, or if no
app is currently fullscreen.
the notifications will now also appear even when fullscreen.