Fixed bugs with Gio's AppInfo.launch
finally ditched the hacky solution i had earlier after fixing the bugs i
had with AppInfo.lauch:
the error that's raised when trying to launch a Terminal=true desktop
entry was fixed by simply linking kitty to ~/.local/bin/xterm, because
gtk is trash and it has a hardcoded list of terminals it looks for, if
none of them is present, it just freaks out and throws the error.
the other error with python applications like lutris was because i'm
using a venv to run fabric-shell, so when launching lutris it'll use the
venv interpreter, and because lutris isn't installed there, it just
crashes. this was solved by creating a gio.launchcontext and stripping
all the venv-specific variables and also cleaning up the PATH, and then
passing the context to the appinfo.launch method.